[codeclim] Refactor to separate ContactCount
