+++ contact_map/__init__.py
@@ -6,8 +6,10 @@ except ImportError:  # pragma: no cover
 __version__ = version.version
 
 from .contact_map import (
-    ContactMap, ContactFrequency, ContactDifference, ContactCount
+    ContactMap, ContactFrequency, ContactDifference
 )
+
+from .contact_count import ContactCount
 
 from .min_dist import NearestAtoms, MinimumDistanceCounter
